[rename cache service]
